==> AllCode/Part_7_NMDS.Rmd <==
---
title: "Part 7: NMDS plots"
author: "Tyler Faits"
date: "12/16/2020"
output: html_document
---

```{r, echo=FALSE, warning=FALSE, message=FALSE}
setwd("/Users/labadmin2/Documents/infant_gill/ProperLabels")
require(ggplot2)
require(ComplexHeatmap)
require(reshape2)
library(RColorBrewer)
require(DESeq2)
require(gridExtra)
require(Rmisc)
require(vegan)
require(circlize)
#require(hclust)
require(NbClust)
#Create color palette to be used for plotting
qual_col_pals = brewer.pal.info[brewer.pal.info$category == 'qual',]
col_vector = unlist(mapply(brewer.pal, qual_col_pals$maxcolors, rownames(qual_col_pals)))

#Load the count matrixes for phylum, genus, and species
phyla <- readRDS("RefSeq2_p.RDS")
genus <- readRDS("RefSeq2_g.RDS")
species <- readRDS("RefSeq2_s.RDS")

#delta/epsilon subdivisions is a part of Proteobacteria, so we're going with that.
phyla["Proteobacteria",] <- phyla["Proteobacteria",] + phyla["delta/epsilon subdivisions",]
phyla <- phyla[rownames(phyla) != "delta/epsilon subdivisions",]
#Require more than 1000 reads in a sample to be analyzed
phyla <- phyla[,apply(phyla,2,sum)>10000]
phyla <- phyla[apply(phyla,1,sum) > 10,]
genus <- genus[,apply(genus,2,sum)>10000]
genus <- genus[apply(genus,1,sum) > 10,]
species <- species[,apply(species,2,sum)>10000]
species <- species[apply(species,1,sum) > 10,]
colnames(genus) <- gsub("S|S0","X",colnames(genus))
colnames(phyla) <- gsub("S|S0","X",colnames(phyla))
colnames(species) <- gsub("S|S0","X",colnames(species))
colnames(genus)[grep("DNA",colnames(genus))] <- c("DNA1","DNA2","DNA3","DNA4","DNA5")
colnames(phyla)[grep("DNA",colnames(phyla))] <- c("DNA1","DNA2","DNA3","DNA4","DNA5")
colnames(species)[grep("DNA",colnames(species))] <- c("DNA1","DNA2","DNA3","DNA4","DNA5")
#Create a counts-per-million matrix for each taxonomic level
CPMphyla <- t(t(phyla)/apply(phyla,2,sum)*1000000)
CPMgenus <- t(t(genus)/apply(genus,2,sum)*1000000)
CPMspecies <- t(t(species)/apply(species,2,sum)*1000000)
map <- readRDS("../InfantMappingData2.RDS")
map$PrePostSympt[is.na(map$PrePostSympt)] <- "none"
map$PrePostSympt[map$PrePostSympt=="none"] <- "Healthy"
map$PrePostSympt[map$Symptoms=="Yes" & !is.na(map$Symptoms)] <- "Active"
map <- map[colnames(genus),]

fullGenus <- readRDS("ggGenusWithClusters.RDS")

BetaDivMap <- vegdist(t(CPMgenus[,map$Sample[map$MothChild=="Infant" & map$Sample %in% colnames(CPMgenus) & map$Age < 1000]]), method="bray")
healthyBetaDivMap <- vegdist(t(CPMgenus[,map$Sample[map$MothChild=="Infant" & map$Sample %in% colnames(CPMgenus) & map$Age < 1000 & map$InfectionStatus=="Healthy"]]), method="bray")
sickBetaDivMap <- vegdist(t(CPMgenus[,map$Sample[map$MothChild=="Infant" & map$Sample %in% colnames(CPMgenus) & map$Age < 1000 & map$InfectionStatus=="RespiratoryIllness"]]), method="bray")
```


```{r, echo=FALSE, warning=FALSE, message=FALSE}
set.seed(5)
preNMDS <- metaMDS(t(CPMgenus[apply(CPMgenus[,map$Sample[map$Age<1000]],1,sum)>0,map$Sample[map$Age<1000]]), distance="bray", k=3,try=10, trymax=500,autotransform = FALSE)
stressplot(preNMDS)
fullGenus$namedCluster6 <- fullGenus$Cluster6
fullGenus$namedCluster6[fullGenus$namedCluster6=="1"] <- "Haemophilus"
fullGenus$namedCluster6[fullGenus$namedCluster6=="2"] <- "Moraxella"
fullGenus$namedCluster6[fullGenus$namedCluster6=="3"] <- "Dolosigranulum"
fullGenus$namedCluster6[fullGenus$namedCluster6=="4"] <- "Anaerobacillus"
fullGenus$namedCluster6[fullGenus$namedCluster6=="5"] <- "Streptococcus"
fullGenus$namedCluster6[fullGenus$namedCluster6=="6"] <- "Staphylococcus"
fullGenus$colors6 <- "#000000"
fullGenus$colors6[fullGenus$namedCluster6=="Haemophilus"] <- col_vector[13]
fullGenus$colors6[fullGenus$namedCluster6=="Moraxella"] <- col_vector[14]
fullGenus$colors6[fullGenus$namedCluster6=="Dolosigranulum"] <- col_vector[12]
fullGenus$colors6[fullGenus$namedCluster6=="Anaerobacillus"] <- col_vector[9]
fullGenus$colors6[fullGenus$namedCluster6=="Streptococcus"] <- col_vector[18]
fullGenus$colors6[fullGenus$namedCluster6=="Staphylococcus"] <- col_vector[17]


DrawInfo <- data.frame(sampNames=map$Sample[map$Age<1000],
                      Colors=unlist(lapply(map$Sample[map$Age<1000],function(x){fullGenus$colors6[fullGenus$Sample==x & fullGenus$Genus=="Staphylococcus"]})),
                      Groups=unlist(lapply(map$Sample[map$Age<1000],function(x){fullGenus$namedCluster6[fullGenus$Sample==x & fullGenus$Genus=="Staphylococcus"]})),stringsAsFactors = FALSE)

taxInfo <- data.frame(Age=map[map$Age<1000,"Age"],Healthy=rep(0,sum(map$Age<1000)),Pre=rep(0,sum(map$Age<1000)),
                         Symptoms=rep(0,sum(map$Age<1000)),Post=rep(0,sum(map$Age<1000)),stringsAsFactors = FALSE)
rownames(taxInfo) <- map$Sample[map$Age<1000]
taxInfo[map[rownames(taxInfo),"InfectionStatus"] == "Healthy","Healthy"] <- 1
taxInfo[map[rownames(taxInfo),"PrePostSympt"] == "Pre","Pre"] <- 1
taxInfo[map[rownames(taxInfo),"Symptoms"] == "Yes" & !is.na(map[rownames(taxInfo),"Symptoms"]),"Symptoms"] <- 1
taxInfo[map[rownames(taxInfo),"PrePostSympt"] == "Post" & taxInfo$Symptoms == 0,"Post"] <- 1

ef <- envfit(preNMDS, taxInfo, permu=999, choices=c(1,2,3))
ordiplot(preNMDS,type="points",display = "sites")
ordiellipse(preNMDS,groups=DrawInfo$Groups, draw="polygon", col=col_vector[c(9,12,13,14,17,18)], label=TRUE)
points(preNMDS, display = 'sites', pch = 19, cex = 0.6, col= DrawInfo$Colors)
plot(ef,col="black",cex=1.5,lwd=4)
lines(c(0,0),c(-1.5,1.5),lty=2)
lines(c(-1.5,2),c(0,0),lty=2)

ef <- envfit(preNMDS, taxInfo, permu=999, choices=c(2,3))
ordiplot(preNMDS,type="points",display = "sites",choices=c(2,3))
ordiellipse(preNMDS,groups=DrawInfo$Groups, draw="polygon", col=col_vector[c(9,12,13,14,17,18)], label=TRUE,choices=c(2,3))
points(preNMDS, display = 'sites', pch = 19, cex = 0.6, col= DrawInfo$Colors,choices=c(2,3))
plot(ef,col="black",cex=1.5,lwd=4)
lines(c(0,0),c(-1.5,1.5),lty=2)
lines(c(-1.5,2),c(0,0),lty=2)

```

```{r, echo=FALSE, warning=FALSE, message=FALSE}
# ordiplot(preNMDS,type="points",display="sites")
# ordiellipse(preNMDS,groups=DrawInfo$Groups,draw="polygon",display="sites",col=col_vector[c(9,12,13,14,17,18)],label=FALSE)
# points(preNMDS, display = 'sites', pch = 19, cex = 0.6, col = DrawInfo$Colors)
# arrows(0,0,bigNMDS$species["Age",1]*0.78,bigNMDS$species["Age",2]*0.78,lwd=2,length=0.15,col="grey18")
# arrows(0,0,bigNMDS$species["Healthy",1]*0.95,bigNMDS$species["Healthy",2]*0.95,lwd=2,length=0.15,col="grey18")
# arrows(0,0,bigNMDS$species["Pre",1]*0.95,bigNMDS$species["Pre",2]*0.95,lwd=2,length=0.15,col="grey18")
# arrows(0,0,bigNMDS$species["Post",1]*0.95,bigNMDS$species["Post",2]*0.95,lwd=2,length=0.15,col="grey18")
# arrows(0,0,bigNMDS$species["Symptoms",1]*0.95,bigNMDS$species["Symptoms",2]*0.95,lwd=2,length=0.15,col="grey18")
# orditorp(bigNMDS,display="species",cex=c(1,1,1,1,1,0.7,0.7,0.7,0.7,0.7,0.7),font=2,priority=rev(c(1,2,3,4,5,6,7,8,9,10,11)))
# #points(bigNMDS$species[c("Strep","Staph","Dolosig","Morax","Haem","Anaero"),],pch=20,cex=4)
# lines(c(0,0),c(-1.5,1.5),lty=2)
# lines(c(-1.5,2),c(0,0),lty=2)
```

```{r, echo=FALSE, warning=FALSE, message=FALSE}
#try the same, but with just one variable for healthy/sick, and one for post (antibiotics)
taxInfo <- data.frame(Age=map[map$Age<1000,"Age"],
                      Sick=rep(0,sum(map$Age<1000)),
                      Healthy=rep(0,sum(map$Age<1000)),
#                      Antibiotic=rep(0,sum(map$Age<1000)),
                      # Staphylococcus=CPMgenus["Staphylococcus",map$Sample[map$Age<1000]],
                      # Streptococcus=CPMgenus["Streptococcus",map$Sample[map$Age<1000]],
                      # Haemophilus=CPMgenus["Haemophilus",map$Sample[map$Age<1000]],
                      # Anaerobacillus=CPMgenus["Anaerobacillus",map$Sample[map$Age<1000]],
                      # Moraxella=CPMgenus["Moraxella",map$Sample[map$Age<1000]],
                      # Dolosigranulum=CPMgenus["Dolosigranulum",map$Sample[map$Age<1000]],
                      stringsAsFactors = FALSE)
rownames(taxInfo) <- map$Sample[map$Age<1000]
taxInfo[map[rownames(taxInfo),"InfectionStatus"] == "RespiratoryIllness","Sick"] <- 1
taxInfo[map[rownames(taxInfo),"InfectionStatus"] == "Healthy","Healthy"] <- 1
#taxInfo[map[rownames(taxInfo),"PrePostSympt"] == "Post","Antibiotic"] <- 1

# taxAndMeta <- data.frame(Age=map[colnames(HUUgenus),"Age"],Sick=rep(0,dim(HUUgenus)[2]),
#                          Antibiotics=rep(0,dim(HUUgenus)[2]),
                         # Staphylococcus=HUUgenus["Staphylococcus",],
                         # Streptococcus=HUUgenus["Streptococcus",],
                         # Haemophilus=HUUgenus["Haemophilus",],
                         # Anaerobacillus=HUUgenus["Anaerobacillus",],
                         # Moraxella=HUUgenus["Moraxella",],
                         # Dolosigranulum=HUUgenus["Dolosigranulum",])

ef <- envfit(preNMDS, choices=c(1,2), taxInfo, permu=999)
ordiplot(preNMDS,type="points",display="sites", choices=c(1,2))
points(preNMDS, display = 'sites', pch = 19, cex = 1, col = DrawInfo$Colors, choices=c(1,2))
ordiellipse(preNMDS,groups=DrawInfo$Groups,draw="polygon",col=col_vector[c(9,12,13,14,17,18)],choices=c(1,2),label=TRUE)
plot(ef,col="black",cex=2,lwd=4)
lines(c(0,0),c(-1.5,1.5),lty=2)
lines(c(-1.5,2),c(0,0),lty=2)

ef <- envfit(preNMDS, choices=c(2,3), taxInfo, permu=999)
ordiplot(preNMDS,type="points",display="sites", choices=c(2,3))
ordiellipse(preNMDS,groups=DrawInfo$Groups,draw="polygon",display="sites",col=col_vector[c(9,12,13,14,17,18)],choices=c(2,3))
points(preNMDS, display = 'sites', pch = 19, cex = 0.6, col = DrawInfo$Colors, choices=c(2,3))
plot(ef,col="black",cex=2,lwd=4)
lines(c(0,0),c(-1.5,1.5),lty=2)
lines(c(-1.5,2),c(0,0),lty=2)
```

```{r}
require(plotly)
myNMDS <- as.data.frame(scores(preNMDS,"sites"))
myNMDS$Colors <- DrawInfo$Colors
myNMDS$Groups <- DrawInfo$Groups
myNMDS$PrePostSympt <- "FAILED"
for(sampName in rownames(myNMDS)){
  myNMDS[sampName,"PrePostSympt"] <- map[sampName,"PrePostSympt"]
}

ef <- envfit(preNMDS, choices=c(1,2,3), taxInfo, permu=999)
ordiplot(preNMDS,type="points",display="sites", choices=c(2,3))
plot_ly(x=myNMDS$NMDS1, y=myNMDS$NMDS2, z=myNMDS$NMDS3,
        type="scatter3d",mode="markers",
        color=DrawInfo$Groups, colors=col_vector[c(28,12,13,14,17,18)]) %>%
  add_trace(x=c(0,ef$vectors$arrows[1,1]),y=c(0,ef$vectors$arrows[1,2]),z=c(0,ef$vectors$arrows[1,3]),
            type = "scatter3d",mode="lines",name="Age",
            line=list(color="black", width=6),
            inherit=FALSE) %>%
  add_trace(x=c(0,ef$vectors$arrows[2,1]),y=c(0,ef$vectors$arrows[2,2]),z=c(0,ef$vectors$arrows[2,3]),
            type = "scatter3d",mode="lines",name="LRTI",
            line=list(color="red", width=6),
            inherit=FALSE) %>%
  add_trace(x=c(0,ef$vectors$arrows[3,1]),y=c(0,ef$vectors$arrows[3,2]),z=c(0,ef$vectors$arrows[3,3]),
            type = "scatter3d",mode="lines",name="Healthy",
            line=list(color="cyan", width=6),
            inherit=FALSE)

ArrowLabels <- data.frame(
  x=c(-0.6,-0.14,0.23),
  y=c(-0.18,-0.28,0.14),
  z=c(0,-0.45,0.45),
  label=c("Age","Healthy","LRTI"),
  Groups=c("Age","Healthy","LRTI")
)
#png("Fig3A.png",961,534)
ggplot(myNMDS, aes(x=NMDS1, y=NMDS2, fill=Groups)) +
    stat_ellipse(level=0.5, geom="polygon", alpha=0.5) +
    stat_ellipse(level=0.5, geom="path") +
    geom_point(size=6, shape=21, color="black") +
    theme_linedraw() +
    scale_fill_manual(name="Cluster membership",values=col_vector[c(28,12,13,14,17,18)]) +
    theme(axis.title = element_text(face="bold", colour="#000000", size=24)) +
    theme(legend.text = element_text(colour="grey35", size = 24, face = "bold.italic")) +
    theme(legend.title = element_text(colour="grey20", size = 28, face = "bold")) + 
    xlab("NMDS axis 1") +
    ylab("NMDS axis 2") +
    theme(axis.text.x  = element_text(size=26, angle = 90, vjust = 0.5)) +
    theme(axis.text.y  = element_text(size=26)) +
    annotate("segment", x=0, y=0, xend=ef$vectors$arrows[1,1], yend=ef$vectors$arrows[1,2],
             arrow=arrow(length=unit(0.7,"cm")), lineend="round", linejoin="mitre",size=2) +
    annotate("segment", x=0, y=0, xend=ef$vectors$arrows[2,1], yend=ef$vectors$arrows[2,2],
             color="red", arrow=arrow(length=unit(0.7,"cm")), lineend="round", linejoin="mitre", size=2) +
    annotate("segment", x=0, y=0, xend=ef$vectors$arrows[3,1], yend=ef$vectors$arrows[3,2],
             color="cyan",arrow=arrow(length=unit(0.7,"cm")), lineend="round", linejoin="mitre", size=2) +
    geom_text(data=ArrowLabels, aes(x=x, y=y, label=label, fill=NA),
              color="black", size=9, angle=c(7,46,46),
              fontface="bold")
#dev.off()

#png("Fig3B.png",961,534)
ggplot(myNMDS, aes(x=NMDS1, y=NMDS2, fill=factor(PrePostSympt,levels=c("Healthy","Pre","Active","Post")))) +
    stat_ellipse(level=0.5, geom="polygon", alpha=0.3) +
    stat_ellipse(level=0.5, geom="path") +
    geom_point(size=6, shape=21, color="black") +
    theme_linedraw() +
    scale_fill_manual(name="LRTI symptom status",values=c("darkred","skyblue","darkgreen","goldenrod"),
                      labels=c("Active LRTI symptoms","Healthy","Post-LRTI","Pre-LRTI")) +
    theme(axis.title = element_text(face="bold", colour="#000000", size=24)) +
    theme(legend.text = element_text(colour="grey35", size = 24, face = "bold.italic")) +
    theme(legend.title = element_text(colour="grey20", size = 28, face = "bold")) + 
    xlab("NMDS axis 1") +
    ylab("NMDS axis 2") +
    theme(axis.text.x  = element_text(size=26, angle = 90, vjust = 0.5)) +
    theme(axis.text.y  = element_text(size=26)) +
    annotate("segment", x=0, y=0, xend=ef$vectors$arrows[1,1], yend=ef$vectors$arrows[1,2],
             arrow=arrow(length=unit(0.7,"cm")), lineend="round", linejoin="mitre",size=2) +
    annotate("segment", x=0, y=0, xend=ef$vectors$arrows[2,1], yend=ef$vectors$arrows[2,2],
             color="red", arrow=arrow(length=unit(0.7,"cm")), lineend="round", linejoin="mitre", size=2) +
    annotate("segment", x=0, y=0, xend=ef$vectors$arrows[3,1], yend=ef$vectors$arrows[3,2],
             color="cyan",arrow=arrow(length=unit(0.7,"cm")), lineend="round", linejoin="mitre", size=2) +
    geom_text(data=ArrowLabels, aes(x=x, y=y, label=label, fill=NA),
              color="black", size=9, angle=c(7,46,46),
              fontface="bold")
#dev.off()

#png("Fig3C.png",961,534)
ggplot(myNMDS, aes(x=NMDS3, y=NMDS2, fill=Groups)) +
    stat_ellipse(level=0.5, geom="polygon", alpha=0.5) +
    stat_ellipse(level=0.5, geom="path") +
    geom_point(size=6, shape=21, color="black") +
    theme_linedraw() +
    scale_fill_manual(name="Cluster membership",values=col_vector[c(28,12,13,14,17,18)]) +
    theme(axis.title = element_text(face="bold", colour="#000000", size=24)) +
    theme(legend.text = element_text(colour="grey35", size = 24, face = "bold.italic")) +
    theme(legend.title = element_text(colour="grey20", size = 28, face = "bold")) + 
    xlab("NMDS axis 3") +
    ylab("NMDS axis 2") +
    theme(axis.text.x  = element_text(size=26, angle = 90, vjust = 0.5)) +
    theme(axis.text.y  = element_text(size=26)) +
    annotate("segment", x=0, y=0, xend=ef$vectors$arrows[1,3], yend=ef$vectors$arrows[1,2],
             arrow=arrow(length=unit(0.7,"cm")), lineend="round", linejoin="mitre",size=2) +
    annotate("segment", x=0, y=0, xend=ef$vectors$arrows[2,3], yend=ef$vectors$arrows[2,2],
             color="red", arrow=arrow(length=unit(0.7,"cm")), lineend="round", linejoin="mitre", size=2) +
    annotate("segment", x=0, y=0, xend=ef$vectors$arrows[3,3], yend=ef$vectors$arrows[3,2],
             color="cyan",arrow=arrow(length=unit(0.7,"cm")), lineend="round", linejoin="mitre", size=2) +
    geom_text(data=ArrowLabels, aes(x=z, y=y, label=label, fill=NA),
              color="black", size=9, angle=c(0,26,26),
              fontface="bold")
#dev.off()

#png("Fig3D.png",961,534)
ggplot(myNMDS, aes(x=NMDS3, y=NMDS2, fill=PrePostSympt)) +
    stat_ellipse(level=0.5, geom="polygon", alpha=0.3) +
    stat_ellipse(level=0.5, geom="path") +
    geom_point(size=6, shape=21, color="black") +
    theme_linedraw() +
    scale_fill_manual(name="LRTI symptom status",values=c("darkred","skyblue","darkgreen","goldenrod"),
                      labels=c("Active LRTI symptoms","Healthy","Post-LRTI","Pre-LRTI")) +
    theme(axis.title = element_text(face="bold", colour="#000000", size=24)) +
    theme(axis.title = element_text(face="bold", colour="#000000", size=24)) +
    theme(legend.text = element_text(colour="grey35", size = 24, face = "bold.italic")) +
    theme(legend.title = element_text(colour="grey20", size = 28, face = "bold")) + 
    xlab("NMDS axis 3") +
    ylab("NMDS axis 2") +
    theme(axis.text.x  = element_text(size=26, angle = 90, vjust = 0.5)) +
    theme(axis.text.y  = element_text(size=26)) +
    annotate("segment", x=0, y=0, xend=ef$vectors$arrows[1,3], yend=ef$vectors$arrows[1,2],
             arrow=arrow(length=unit(0.7,"cm")), lineend="round", linejoin="mitre",size=2) +
    annotate("segment", x=0, y=0, xend=ef$vectors$arrows[2,3], yend=ef$vectors$arrows[2,2],
             color="red", arrow=arrow(length=unit(0.7,"cm")), lineend="round", linejoin="mitre", size=2) +
    annotate("segment", x=0, y=0, xend=ef$vectors$arrows[3,3], yend=ef$vectors$arrows[3,2],
             color="cyan",arrow=arrow(length=unit(0.7,"cm")), lineend="round", linejoin="mitre", size=2) +
    geom_text(data=ArrowLabels, aes(x=z, y=y, label=label, fill=NA),
              color="black", size=9, angle=c(0,26,26),
              fontface="bold")
#dev.off()
  
```


```{r}
# complexNMDS <- metaMDS(t(CPMgenus[apply(CPMgenus[,map$Sample[map$Age<1000]],1,sum)>0,map$Sample[map$Age<1000]]), distance="bray", k=4,try=50, trymax=100,autotransform = FALSE)
# stressplot(complexNMDS)
# 
# ef <- envfit(complexNMDS, choices=c(1,2,3,4), taxInfo, permu=999)
# mycompNMDS <- as.data.frame(scores(complexNMDS,"sites"))
# 
# plot_ly(x=mycompNMDS$NMDS1, y=mycompNMDS$NMDS2, z=mycompNMDS$NMDS3,
#         type="scatter3d",mode="markers",
#         color=DrawInfo$Groups, colors=col_vector[c(9,12,13,14,17,18)]) %>%
#   add_trace(x=c(0,-0.98191),y=c(0,-0.12104),z=c(0,0.13933),
#             type = "scatter3d",mode="lines",name="Age",
#             line=list(color="black", width=6),
#             inherit=FALSE) %>%
#   add_trace(x=c(0,0.31531),y=c(0,0.42872),z=c(0,-0.72352),
#             type = "scatter3d",mode="lines",name="LRTI",
#             line=list(color="red", width=6),
#             inherit=FALSE) %>%
#   add_trace(x=c(0,-0.31531),y=c(0,-0.42872),z=c(0,0.72352),
#             type = "scatter3d",mode="lines",name="Healthy",
#             line=list(color="blue", width=6),
#             inherit=FALSE)
```
